model/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
Pillow
python-dotenv
python-multipart
psycopg2-binary